requirements.txt
# filepath: /home/ubuntu/agentic-pipeline/requirements.txt

langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.2.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
requests>=2.32.0
pytest>=8.0.0
